make mesh cache one of the generic caches
