+++ crates/re_viewer/src/misc/caches/mod.rs
@@ -1,98 +1,9 @@
+mod caches;
 mod mesh_cache;
 mod tensor_decode_cache;
 
-use ahash::HashMap;
-use re_log_types::component_types;
-
-/// Does memoization of different things for the immediate mode UI.
-#[derive(Default)]
-pub struct Caches {
-    /// Cached decoded tensors.
-    pub decode: tensor_decode_cache::DecodeCache,
-
-    /// Cached loaded meshes (from file or converted from user data).
-    pub mesh: mesh_cache::MeshCache,
-
-    tensor_stats: nohash_hasher::IntMap<component_types::TensorId, TensorStats>,
-
-    caches: HashMap<std::any::TypeId, Box<dyn Cache + 'static>>,
-}
-
-impl Caches {
-    /// Call once per frame to potentially flush the cache(s).
-    pub fn begin_frame(&mut self) {
-        #[cfg(not(target_arch = "wasm32"))]
-        let max_decode_cache_use = 4_000_000_000;
-
-        #[cfg(target_arch = "wasm32")]
-        let max_decode_cache_use = 1_000_000_000;
-
-        self.decode.begin_frame(max_decode_cache_use);
-
-        for cache in self.caches.values_mut() {
-            cache.begin_frame();
-        }
-    }
-
-    pub fn purge_memory(&mut self) {
-        let Self {
-            decode,
-            tensor_stats,
-            mesh: _, // TODO(emilk)
-            caches,
-        } = self;
-        decode.purge_memory();
-        tensor_stats.clear();
-
-        for cache in caches.values_mut() {
-            cache.purge_memory();
-        }
-    }
-
-    pub fn tensor_stats(&mut self, tensor: &re_log_types::component_types::Tensor) -> &TensorStats {
-        self.tensor_stats
-            .entry(tensor.tensor_id)
-            .or_insert_with(|| TensorStats::new(tensor))
-    }
-
-    /// Retrieves a cache for reading and writing.
-    ///
-    /// Returns None if the cache is not present.
-    pub fn get_mut<T: Cache>(&mut self) -> Option<&mut T> {
-        self.caches
-            .get_mut(&std::any::TypeId::of::<T>())
-            .and_then(|cache| <dyn std::any::Any>::downcast_mut::<T>(cache))
-    }
-
-    /// Retrieves a cache for reading.
-    ///
-    /// Returns None if the cache is not present.
-    pub fn get<T: Cache>(&self) -> Option<&T> {
-        self.caches
-            .get(&std::any::TypeId::of::<T>())
-            .and_then(|cache| <dyn std::any::Any>::downcast_ref::<T>(cache))
-    }
-
-    /// Adds a cache to the list of caches.
-    ///
-    /// Fails if a cache of the same type already exists.
-    pub fn add_cache<T: Cache>(&mut self, cache: T) -> Result<(), ()> {
-        let type_id = std::any::TypeId::of::<T>();
-        match self.caches.insert(type_id, Box::new(cache)) {
-            Some(_) => Err(()),
-            None => Ok(()),
-        }
-    }
-}
-
-/// A cache for memoizing things in order to speed up immediate mode UI & other immediate mode style things.
-pub trait Cache: std::any::Any {
-    /// Called once per frame to potentially flush the cache.
-    fn begin_frame(&mut self);
-
-    /// Attempt to free up memory.
-    fn purge_memory(&mut self);
-}
+pub use caches::{Cache, Caches};
+pub use mesh_cache::MeshCache;
 
 #[derive(Clone, Copy, Debug)]
 pub struct TensorStats {
